Reverting to fill answer array with entire question object as key.
I had changed it to use the question slug as the key which improved how previously answered questions were filled in after a page refresh. But this approach left previous answers blank if having not refreshed the browser and instead just navigate back. It also caused skip logic to no longer work. Might return to the question slug approach but with more thorough code changes.